The Bible Verse Memory App/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
ics==0.7
python-dateutil==2.8.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.100.0
resend==3.0.0
python-dotenv==1.0.0gunicorn==21.2.0